+++ CHANGELOG.md
@@ -1,3 +1,3 @@
-## v1.0.0
-+ Added Semantic Versioning
-+ Added versionctl helper tool
+## v1.1.0
++ Changed gpu stat to use glxinfo instead of lspci
++ Added integrated and dedicated GPU detection